(PC-29071)[PRO] refactor: move mapDayToFrench to date utils
