+++ pro/src/pages/VenueEdition/OpeningHoursReadOnly/OpeningHoursReadOnly.tsx
@@ -1,9 +1,9 @@
 import { GetVenueResponseModel } from 'apiClient/v1'
 import { SummaryDescriptionList } from 'components/SummaryLayout/SummaryDescriptionList'
 import { SummarySubSection } from 'components/SummaryLayout/SummarySubSection'
+import { mapDayToFrench } from 'utils/date'
 
 import styles from './OpeningHoursReadOnly.module.scss'
-import { mapDayToFrench } from './utils'
 
 type OpeningHours = {
   openingHours: GetVenueResponseModel['openingHours']
