+++ pro/src/utils/date.ts
@@ -114,3 +114,33 @@ export function formatTimeForInput(date: Date) {
     minute: '2-digit',
   }).format(date)
 }
+
+export function mapDayToFrench(
+  day: string
+):
+  | 'Lundi'
+  | 'Mardi'
+  | 'Mercredi'
+  | 'Jeudi'
+  | 'Vendredi'
+  | 'Samedi'
+  | 'Dimanche' {
+  switch (day.toUpperCase()) {
+    case 'MONDAY':
+      return 'Lundi'
+    case 'TUESDAY':
+      return 'Mardi'
+    case 'WEDNESDAY':
+      return 'Mercredi'
+    case 'THURSDAY':
+      return 'Jeudi'
+    case 'FRIDAY':
+      return 'Vendredi'
+    case 'SATURDAY':
+      return 'Samedi'
+    case 'SUNDAY':
+      return 'Dimanche'
+    default:
+      return 'Dimanche'
+  }
+}
